Added document comments to CameraServo class.

+++ CameraServo.cpp
@@ -23,6 +23,18 @@ class CameraServo {
 	
 	public:
 	
+		/*
+		 * Creates a new CameraServo instance and configures constants.
+		 * 
+		 * @param int PWM port the pan servo is on.
+		 * @param int PWM port the tilt servo is on.
+		 * @param float Value between 0.0 and 1.0 where the pan servo is 
+		 * centered.
+		 * @param float Value between 0.0 and 1.0 where the tilt servo is 
+		 * centered.
+		 * @param float Specify which value is the left direction (-1.0 or 1.0).
+		 * @param float Specify which value is the up direction (-1.0 or 1.0).
+		 */
 		CameraServo(int panPort, int tiltPort, float panCenter, float tiltCenter, float panLeftDirection, float tiltUpDirection):
 			pan(panPort),
 			tilt(tiltPort)
@@ -30,19 +42,29 @@ class CameraServo {
 			CAMERA_PAN_CENTER  = panCenter;
 			CAMERA_TILT_CENTER = tiltCenter;
 			
+			// Configurable directions.
 			PAN_LEFT_MULTIPLIER    = panLeftDirection;
 			PAN_RIGHT_MULTIPLIER   = panLeftDirection * -1;
 			TILT_UP_MULTIPLIER     = tiltUpDirection;
 			TILT_DOWN_MULTIPLIER   = tiltUpDirection * -1;
 		}
 		
-
+		/*
+		 * Sets the camera to it's center position.
+		 */
 		void SetToCenter() {
 			pan.Set(CAMERA_PAN_CENTER);
 			tilt.Set(CAMERA_TILT_CENTER);
 		}
 		
-		bool PanLeft(int positionOffset = 0.05) {
+		/*
+		 * Move the camera to the left.
+		 * 
+		 * @param float Value between 0.0 and 1.0 of the step the pan motor will 
+		 * take to the left.  Default is 0.05.  It would take 20 steps to go 
+		 * from 0.0 to 1.0.
+		 */
+		bool PanLeft(float positionOffset = 0.05) {
 			float currentPosition = pan.Get();
 			if (currentPosition == 0.0)
 				return false;
@@ -51,7 +73,14 @@ class CameraServo {
 			return true;
 		}
 		
-		bool PanRight(int positionOffset = 0.05) {
+		/*
+		 * Move the camera to the right.
+		 * 
+		 * @param float Value between 0.0 and 1.0 of the step the pan motor will
+		 * take to the right.  Default is 0.05.  It would take 20 steps to go
+		 * from 0.0 to 1.0.
+		 */
+		bool PanRight(float positionOffset = 0.05) {
 			float currentPosition = pan.Get();
 			if (currentPosition == 1.0)
 				return false;
@@ -60,7 +89,14 @@ class CameraServo {
 			return true;
 		}
 		
-		bool TiltUp(int positionOffset = 0.05) {
+		/*
+		 * Tilt the camera up.
+		 * 
+		 * @param float Value between 0.0 to 1.0 of the step the tilt motor will
+		 * tilt up by.  Default is 0.05.  It would take 20 steps to go from 0.0
+		 * to 1.0.
+		 */
+		bool TiltUp(float positionOffset = 0.05) {
 			float currentPosition = tilt.Get();
 			if (currentPosition == 0.0)
 				return false;
@@ -69,7 +105,14 @@ class CameraServo {
 			return true; 
 		}
 		
-		bool TiltDown(int positionOffset = 0.05) {
+		/*
+		 * Tilt the camera down.
+		 * 
+		 * @param float Value between 0.0 to 1.0 of the step the tilt motor will
+		 * tilt down by.  Default is 0.05.  It would take 20 steps to go from 
+		 * 0.0 to 1.0.
+		 */
+		bool TiltDown(float positionOffset = 0.05) {
 			float currentPosition = tilt.Get();
 			if (currentPosition == 1.0)
 				return true;
